refactored test cases for expense manager
